Fixes DE216 - glance-registry-paste.ini password
    [This could be separated into 2 commits]

The main organizational change here is removing the big chunk of code
from glance::registry that searches the node index for glance::setup to
find the db/admin passwords.  This unnecessarily duplicates
RCB#get_settings_by_role, which is called in glance::common anyway
(*BEFORE* this code even runs).  That's the first bug--the templates in
glance::common then try to use node values (set by this code) which
haven't been assigned yet.  So instead of fixing the ordering of the
code, just delete this unnecessary block and use the values discovered
by get_settings_by_role().

The secondary changes are stylistic cleanup for readability and maint.