Change dsh group name back to 'glance'
Originally, it didn't look like you could update an existing dsh group so the
idea was to define a new one for glance user.  However, there was a
misunderstanding what the group attribute did but fortunately a bug existed
in the dsh cookbook where it'd never get used.  That bug has since been fixed
and it also appears that dsh groups can now be updated, so we just revert to
calling it glance.