Change the AMQP queue when using file backend
These changes make it possible to have glance image sync function
when using a backend store of "file".

When a backend store of "file" is used the glance notification topic
queue will be set to "glance_notifications"; this will override the
"default" attribute.

These change work in conjuction with the glance image sync script:
https://github.com/rcbops/glance-image-sync